--- a/src/runtime/gc_blocks.go
+++ b/src/runtime/gc_blocks.go
@@ -534,6 +534,12 @@ func markRoots(start, end uintptr) {
 	}
 }
 
+func markCurrentGoroutineStack(sp uintptr) {
+	// This could be optimized by only marking the stack area that's currently
+	// in use.
+	markRoot(0, sp)
+}
+
 // stackOverflow is a flag which is set when the GC scans too deep while marking.
 // After it is set, all marked allocations must be re-scanned.
 var stackOverflow bool