@pulsejet 2025-03-10
Is it currently planned to finish this work? We're using tinygo for wasm because using asyncify is a huge win for optimization and binary size, but the gc bottleneck makes it unusable (pauses are ~400ms vs ~20ms for gc)