@aykevl 2025-03-14
Updated the PR, I think it's still valuable to have.