@deadprogram 2025-03-14
> Updated the PR, I think it's still valuable to have.

Yes, please!

> In the future, this GC might also be supported on WebAssembly with some extra work. 

IMO this is a very important feature to add.