@aykevl 2025-03-14
Here is a comparison benchmarking the html package:

```
$ benchstat html-precise.txt html-boehm.txt
               │ html-precise.txt │            html-boehm.txt             │
               │      sec/op      │    sec/op     vs base                 │
Escape               23.68µ ± ∞ ¹   16.52µ ± ∞ ¹        ~ (p=1.000 n=1) ²
EscapeNone           9.007µ ± ∞ ¹   8.990µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Unescape             41.26µ ± ∞ ¹   27.56µ ± ∞ ¹        ~ (p=1.000 n=1) ²
UnescapeNone         1.885µ ± ∞ ¹   1.946µ ± ∞ ¹        ~ (p=1.000 n=1) ²
UnescapeSparse      10.465µ ± ∞ ¹   6.265µ ± ∞ ¹        ~ (p=1.000 n=1) ²
UnescapeDense        35.62µ ± ∞ ¹   25.57µ ± ∞ ¹        ~ (p=1.000 n=1) ²
geomean              13.55µ         10.41µ        -23.13%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05
```

If I read this correctly, it's 23% faster.

And for go/format, with `-gc=boehm`:

```
BenchmarkFormat/array1-10000                 402          14963832 ns/op           4.33 MB/s     7324959 B/op          0 allocs/op
```

`-gc=precise`:

```
BenchmarkFormat/array1-10000                 144          63817734 ns/op           1.01 MB/s     8169418 B/op     204348 allocs/op
```

(The 0 allocs/op are likely incorrect, but the ns/op is massively faster).