@aykevl 2025-03-14
Ran a few more benchmarks, this time with the `-count` parameter so benchstat can actually run some statistics on the numbers.

go/format benchmark, which seems to be 321% faster (or take 76% less time):

```
$ benchstat go-format-precise.txt go-format-boehm.txt
                    │ go-format-precise.txt │         go-format-boehm.txt         │
                    │        sec/op         │   sec/op     vs base                │
Format/array1-10000            63.53m ± 43%   15.06m ± 1%  -76.30% (p=0.000 n=10)

                    │ go-format-precise.txt │          go-format-boehm.txt           │
                    │          B/s          │      B/s       vs base                 │
Format/array1-10000           996.1Ki ± 31%   4199.2Ki ± 1%  +321.57% (p=0.000 n=10)

                    │ go-format-precise.txt │         go-format-boehm.txt          │
                    │         B/op          │     B/op      vs base                │
Format/array1-10000            7.790Mi ± 0%   6.980Mi ± 0%  -10.40% (p=0.000 n=10)
```

html benchmark, which seems to take 21% less time:

```
$ benchstat html-precise.txt html-boehm.txt
               │ html-precise.txt │           html-boehm.txt            │
               │      sec/op      │   sec/op     vs base                │
Escape                24.39µ ± 0%   17.78µ ± 2%  -27.08% (p=0.000 n=10)
EscapeNone            9.370µ ± 1%   9.329µ ± 2%        ~ (p=0.172 n=10)
Unescape              42.21µ ± 1%   29.53µ ± 1%  -30.03% (p=0.000 n=10)
UnescapeNone          1.922µ ± 2%   1.968µ ± 1%   +2.42% (p=0.000 n=10)
UnescapeSparse       10.489µ ± 3%   6.529µ ± 0%  -37.75% (p=0.000 n=10)
UnescapeDense         36.62µ ± 1%   26.69µ ± 2%  -27.10% (p=0.000 n=10)
geomean               13.87µ        10.90µ       -21.38%
```

crypto/sha1 might be a little bit slower (not sure why):

```
$ benchstat sha1-precise.txt sha1-boehm.txt
                 │ sha1-precise.txt │           sha1-boehm.txt           │
                 │      sec/op      │   sec/op     vs base               │
Hash8Bytes/New          178.3n ± 1%   180.2n ± 1%  +1.09% (p=0.019 n=10)
Hash8Bytes/Sum          170.0n ± 1%   171.6n ± 1%  +0.91% (p=0.005 n=10)
Hash320Bytes/New        949.8n ± 1%   947.8n ± 2%       ~ (p=0.796 n=10)
Hash320Bytes/Sum        944.3n ± 1%   947.4n ± 2%       ~ (p=0.247 n=10)
Hash1K/New              2.679µ ± 1%   2.682µ ± 1%       ~ (p=0.286 n=10)
Hash1K/Sum              2.673µ ± 1%   2.694µ ± 1%  +0.77% (p=0.015 n=10)
Hash8K/New              19.96µ ± 1%   20.29µ ± 1%  +1.65% (p=0.000 n=10)
Hash8K/Sum              20.08µ ± 1%   20.34µ ± 1%  +1.32% (p=0.007 n=10)
geomean                 1.724µ        1.737µ       +0.75%

                 │ sha1-precise.txt │           sha1-boehm.txt            │
                 │       B/s        │     B/s       vs base               │
Hash8Bytes/New         42.79Mi ± 1%   42.33Mi ± 1%  -1.06% (p=0.022 n=10)
Hash8Bytes/Sum         44.88Mi ± 1%   44.47Mi ± 1%  -0.89% (p=0.007 n=10)
Hash320Bytes/New       321.3Mi ± 1%   322.0Mi ± 2%       ~ (p=0.796 n=10)
Hash320Bytes/Sum       323.2Mi ± 1%   322.1Mi ± 2%       ~ (p=0.247 n=10)
Hash1K/New             364.6Mi ± 1%   364.1Mi ± 1%       ~ (p=0.280 n=10)
Hash1K/Sum             365.3Mi ± 1%   362.6Mi ± 1%  -0.75% (p=0.015 n=10)
Hash8K/New             391.4Mi ± 1%   385.0Mi ± 1%  -1.63% (p=0.000 n=10)
Hash8K/Sum             389.1Mi ± 1%   384.0Mi ± 1%  -1.31% (p=0.007 n=10)
geomean                211.8Mi        210.2Mi       -0.74%
```

encoding/hex is a lot faster in some benchmarks:

```
$ benchstat hex-precise.txt hex-boehm.txt 
                   │ hex-precise.txt │            hex-boehm.txt            │
                   │     sec/op      │    sec/op     vs base               │
Encode/256              202.3n ± ∞ ¹   204.0n ± ∞ ¹        ~ (p=0.257 n=4)
Encode/1024             782.2n ± ∞ ¹   794.3n ± ∞ ¹   +1.55% (p=0.029 n=4)
Encode/4096             3.171µ ± ∞ ¹   3.184µ ± ∞ ¹        ~ (p=0.686 n=4)
Encode/16384            12.51µ ± ∞ ¹   12.79µ ± ∞ ¹   +2.25% (p=0.029 n=4)
Decode/256              136.5n ± ∞ ¹   138.6n ± ∞ ¹        ~ (p=0.114 n=4)
Decode/1024             520.8n ± ∞ ¹   528.8n ± ∞ ¹        ~ (p=0.200 n=4)
Decode/4096             2.081µ ± ∞ ¹   2.102µ ± ∞ ¹        ~ (p=0.343 n=4)
Decode/16384            8.354µ ± ∞ ¹   8.443µ ± ∞ ¹   +1.07% (p=0.029 n=4)
DecodeString/256        243.2n ± ∞ ¹   187.1n ± ∞ ¹  -23.07% (p=0.029 n=4)
DecodeString/1024       918.4n ± ∞ ¹   692.4n ± ∞ ¹  -24.61% (p=0.029 n=4)
DecodeString/4096       3.826µ ± ∞ ¹   3.181µ ± ∞ ¹  -16.85% (p=0.029 n=4)
DecodeString/16384      20.77µ ± ∞ ¹   11.84µ ± ∞ ¹  -43.00% (p=0.029 n=4)
Dump/256                5.910µ ± ∞ ¹   5.799µ ± ∞ ¹   -1.89% (p=0.029 n=4)
Dump/1024               25.18µ ± ∞ ¹   21.27µ ± ∞ ¹  -15.54% (p=0.029 n=4)
Dump/4096              102.51µ ± ∞ ¹   91.09µ ± ∞ ¹  -11.14% (p=0.029 n=4)
Dump/16384              397.1µ ± ∞ ¹   334.5µ ± ∞ ¹  -15.77% (p=0.029 n=4)
geomean                 3.611µ         3.253µ         -9.90%
¹ need >= 6 samples for confidence interval at level 0.95

                   │ hex-precise.txt │             hex-boehm.txt             │
                   │       B/s       │      B/s        vs base               │
Encode/256             1.179Gi ± ∞ ¹    1.169Gi ± ∞ ¹        ~ (p=0.200 n=4)
Encode/1024            1.219Gi ± ∞ ¹    1.201Gi ± ∞ ¹   -1.53% (p=0.029 n=4)
Encode/4096            1.203Gi ± ∞ ¹    1.198Gi ± ∞ ¹        ~ (p=0.686 n=4)
Encode/16384           1.220Gi ± ∞ ¹    1.193Gi ± ∞ ¹   -2.20% (p=0.029 n=4)
Decode/256             1.746Gi ± ∞ ¹    1.720Gi ± ∞ ¹        ~ (p=0.114 n=4)
Decode/1024            1.831Gi ± ∞ ¹    1.804Gi ± ∞ ¹        ~ (p=0.200 n=4)
Decode/4096            1.833Gi ± ∞ ¹    1.815Gi ± ∞ ¹        ~ (p=0.343 n=4)
Decode/16384           1.827Gi ± ∞ ¹    1.807Gi ± ∞ ¹   -1.05% (p=0.029 n=4)
DecodeString/256      1003.9Mi ± ∞ ¹   1304.9Mi ± ∞ ¹  +29.98% (p=0.029 n=4)
DecodeString/1024      1.038Gi ± ∞ ¹    1.377Gi ± ∞ ¹  +32.64% (p=0.029 n=4)
DecodeString/4096     1021.1Mi ± ∞ ¹   1227.9Mi ± ∞ ¹  +20.26% (p=0.029 n=4)
DecodeString/16384     753.5Mi ± ∞ ¹   1319.5Mi ± ∞ ¹  +75.11% (p=0.029 n=4)
Dump/256               41.31Mi ± ∞ ¹    42.10Mi ± ∞ ¹   +1.93% (p=0.029 n=4)
Dump/1024              38.79Mi ± ∞ ¹    45.92Mi ± ∞ ¹  +18.40% (p=0.029 n=4)
Dump/4096              38.11Mi ± ∞ ¹    42.88Mi ± ∞ ¹  +12.53% (p=0.029 n=4)
Dump/16384             39.34Mi ± ∞ ¹    46.71Mi ± ∞ ¹  +18.73% (p=0.029 n=4)
geomean                541.0Mi          600.4Mi        +10.98%
¹ need >= 6 samples for confidence interval at level 0.95
```