@aykevl 2025-03-17
I did a quick-and-dirty port to wasip1, benchmarked the html package, and got the following result:

```
$ benchstat html-precise.txt html-boehm.txt
               │ html-precise.txt │           html-boehm.txt           │
               │      sec/op      │   sec/op     vs base               │
Escape               110.29µ ± 0%   73.62µ ± 3%  -33.25% (p=0.002 n=6)
EscapeNone            48.37µ ± 0%   49.15µ ± 1%   +1.62% (p=0.002 n=6)
Unescape             192.31µ ± 5%   84.53µ ± 3%  -56.04% (p=0.002 n=6)
UnescapeNone          2.638µ ± 1%   2.559µ ± 4%        ~ (p=0.132 n=6)
UnescapeSparse        45.44µ ± 2%   18.55µ ± 1%  -59.17% (p=0.002 n=6)
UnescapeDense        182.55µ ± 1%   79.63µ ± 1%  -56.38% (p=0.002 n=6)
geomean               53.11µ        32.40µ       -39.00%
```

A geomean of -39% is very significant! It's also a lot bigger than the difference on Linux.
(GC pauses are probably also reduced, but I didn't check those).

I also tried benchmarking go/format on wasip1 and while that one did pass with `-gc=boehm`, it actually crashed with `-gc=precise` (out of memory). So `-gc=boehm` is an improvement there as well. It also seemed to improve performance a lot.